Fixes Current File Panel report of not found keys
Derived keys (e.g., such as plurals) are being miscategorised as not
found in the Current File Panel report.  This commit checks if any
existing key paths match derived keys paths, and if so, includes them in
the list of in use key paths and removed from the "not found" list.

Fixes https://github.com/lokalise/i18n-ally/issues/959