feat: add `i18next-shopify` framework (#970)
* Return first available counting value if nested

* Add extraction support to react-i18next framework

This commit adds the `supportAutoExtraction` array of supported formats
and the `detectHardStrings` method from `GeneralFramework` to enable
hard-coded string extraction for react-i18next framework, without
needing the "general" framework to be enabled/loaded in the
`.vscode/settings.json` file.

* Add i18next-shopify framework

This commit adds the i18next-shopify framework to the list of supported
frameworks. It is an extension of the React i18next framework, with some
configuration changes.

Also add i18next-shopify example app, including a pluralization example.

* Add support for Go to Definition of a translation key

* Fixes usage report for non-missing derived keys

Derived keys (e.g., such as plurals) were being miscategorised as
missing in the usage report.  This commit checks if keys derived from
those found in the missing list (keys that are in use but considered
"not defined") are included in the list of idle keys (keys defined but
not considered in use), indicating that they are indeed defined and in
use.

Fixes https://github.com/lokalise/i18n-ally/issues/953

* Fixes Current File Panel report of not found keys

Derived keys (e.g., such as plurals) are being miscategorised as not
found in the Current File Panel report.  This commit checks if any
existing key paths match derived keys paths, and if so, includes them in
the list of in use key paths and removed from the "not found" list.

Fixes https://github.com/lokalise/i18n-ally/issues/959

* Chore: fix spellings in code base

---------

Co-authored-by: Kevin O'Sullivan <mkevin.osullivan@shopify.com>